Do not recommend to rely on hashes being ordered
Although an ordered hash can come in useful at times, hashes in general
are often thought of as begin unordered. An implementation that relies
on the fact that a hash retains the order of its keys can therefore
possibly be a source of confusion. We should not explicitly recommend
it. Let developer use brain and decide case by case.